working on making all adapter tests share common test script
