+++ packages/datastore-storage-adapter/__tests__/helpers.ts
@@ -3,7 +3,6 @@ import {
 	MutableModel,
 	Schema,
 	InternalSchema,
-	SchemaModel,
 } from '@aws-amplify/datastore';
 
 export declare class Model {
