+++ packages/datastore-storage-adapter/package.json
@@ -50,6 +50,7 @@
 					"lib": [
 						"es5",
 						"es2015",
+						"dom",
 						"esnext.asynciterable",
 						"es2019"
 					],
@@ -89,7 +90,8 @@
 			"/node_modules/",
 			"dist",
 			"lib",
-			"lib-esm"
+			"lib-esm",
+			"../datastore"
 		]
 	}
 }
