+++ packages/datastore/__tests__/IndexedDBAdapter.test.ts
@@ -5,87 +5,38 @@ import {
 	initSchema as initSchemaType,
 } from '../src/datastore/datastore';
 import { PersistentModelConstructor, SortDirection } from '../src/types';
-import { Model, User, Profile, Post, Comment, testSchema } from './helpers';
+import {
+	pause,
+	expectMutation,
+	Model,
+	User,
+	Profile,
+	Post,
+	Comment,
+	testSchema,
+} from './helpers';
 import { Predicates } from '../src/predicates';
+import { addCommonQueryTests } from './commonAdapterTests';
 
 let initSchema: typeof initSchemaType;
 let DataStore: typeof DataStoreType;
 // using any to get access to private methods
 const IDBAdapter = <any>Adapter;
 
-async function pause(ms) {
-	return new Promise(resolve => setTimeout(resolve, ms));
-}
-
-const UUID_REGEX =
-	/^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/;
-
-/**
- * Tests a mutation for expected values. If values are present on the mutation
- * that are not expected, throws an error. Expected values can be listed as a
- * literal value, a regular expression, or a function (v => bool).
- *
- * `id` is automatically tested and expected to be a UUID unless an alternative
- *  matcher is provided.
- *
- * @param mutation A mutation record to check
- * @param values An object for specific values to test. Format of key: value | regex | v => bool
- */
-function expectMutation(mutation, values) {
-	const data = JSON.parse(mutation.data);
-	const matchers = {
-		id: UUID_REGEX,
-		...values,
-	};
-	const errors = [
-		...errorsFrom(data, matchers),
-		...extraFieldsFrom(data, matchers).map(f => `Unexpected field: ${f}`),
-	];
-	if (errors.length > 0) {
-		throw new Error(
-			`Bad mutation: ${JSON.stringify(data, null, 2)}\n${errors.join('\n')}`
-		);
-	}
-}
-
-/**
- * Checks an object for adherence to expected values from a set of matchers.
- * Returns a list of erroneous key-value pairs.
- * @param data the object to validate.
- * @param matchers the matcher functions/values/regexes to test the object with
- */
-function errorsFrom(data, matchers) {
-	return Object.entries(matchers).reduce((errors, [property, matcher]) => {
-		const value = data[property];
-		if (
-			!(
-				(typeof matcher === 'function' && matcher(value)) ||
-				(matcher instanceof RegExp && matcher.test(value)) ||
-				value === matcher
-			)
-		) {
-			errors.push(
-				`Property '${property}' value "${value}" does not match "${matcher}"`
-			);
-		}
-		return errors;
-	}, []);
-}
-
-/**
- * Checks to see if a given object contains any extra, unexpected properties.
- * If any are present, it returns the list of unexpectd fields.
- *
- * @param data the object that MIGHT contain extra fields.
- * @param template the authorative template object.
- */
-function extraFieldsFrom(data, template) {
-	const fields = Object.keys(data);
-	const expectedFields = new Set(Object.keys(template));
-	return fields.filter(name => !expectedFields.has(name));
-}
-
 describe('IndexedDBAdapter tests', () => {
+	async function getMutations(adapter) {
+		await pause(250);
+		return await adapter.getAll('sync_MutationEvent');
+	}
+
+	({ initSchema, DataStore } = require('../src/datastore/datastore'));
+	addCommonQueryTests({
+		initSchema,
+		DataStore,
+		storageAdapter: Adapter,
+		getMutations,
+	});
+
 	describe('Query', () => {
 		let Model: PersistentModelConstructor<Model>;
 		let model1Id: string;
@@ -97,6 +48,7 @@ describe('IndexedDBAdapter tests', () => {
 
 		beforeAll(async () => {
 			({ initSchema, DataStore } = require('../src/datastore/datastore'));
+			DataStore.configure({ storageAdapter: Adapter });
 
 			const classes = initSchema(testSchema());
 
@@ -110,6 +62,9 @@ describe('IndexedDBAdapter tests', () => {
 			// quickly enough that dates were colliding. (or so it seemed!)
 
 			const baseDate = new Date();
+
+			await DataStore.start();
+			await DataStore.clear();
 
 			({ id: model1Id } = await DataStore.save(
 				new Model({
@@ -187,120 +142,6 @@ describe('IndexedDBAdapter tests', () => {
 			expect(spyOnEngine).not.toHaveBeenCalled();
 			expect(spyOnMemory).not.toHaveBeenCalled();
 		});
-
-		it('should match fields of any non-empty value for `("ne", undefined)`', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('ne', undefined)
-			);
-			expect(results.length).toEqual(3);
-		});
-
-		it('should match fields of any non-empty value for `("ne", null)`', async () => {
-			const results = await DataStore.query(Model, m => m.field1('ne', null));
-			expect(results.length).toEqual(3);
-		});
-
-		it('should NOT match fields of any non-empty value for `("eq", undefined)`', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('eq', undefined)
-			);
-			expect(results.length).toEqual(0);
-		});
-
-		it('should NOT match fields of any non-empty value for `("eq", null)`', async () => {
-			const results = await DataStore.query(Model, m => m.field1('eq', null));
-			expect(results.length).toEqual(0);
-		});
-
-		it('should NOT match fields of any non-empty value for `("gt", null)`', async () => {
-			const results = await DataStore.query(Model, m => m.field1('gt', null));
-			expect(results.length).toEqual(0);
-		});
-
-		it('should NOT  match fields of any non-empty value for `("ge", null)`', async () => {
-			const results = await DataStore.query(Model, m => m.field1('ge', null));
-			expect(results.length).toEqual(0);
-		});
-
-		it('should NOT match fields of any non-empty value for `("lt", null)`', async () => {
-			const results = await DataStore.query(Model, m => m.field1('lt', null));
-			expect(results.length).toEqual(0);
-		});
-
-		it('should NOT match fields of any non-empty value for `("le", null)`', async () => {
-			const results = await DataStore.query(Model, m => m.field1('le', null));
-			expect(results.length).toEqual(0);
-		});
-
-		it('should NOT match fields of any non-empty value for `("gt", undefined)`', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('gt', undefined)
-			);
-			expect(results.length).toEqual(0);
-		});
-
-		it('should NOT match fields of any non-empty value for `("ge", undefined)`', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('ge', undefined)
-			);
-			expect(results.length).toEqual(0);
-		});
-
-		it('should NOT match fields of any non-empty value for `("lt", undefined)`', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('lt', undefined)
-			);
-			expect(results.length).toEqual(0);
-		});
-
-		it('should NOT match fields of any non-empty value for `("le", undefined)`', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('le', undefined)
-			);
-			expect(results.length).toEqual(0);
-		});
-
-		it('should match gt', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('gt', 'field1 value 0')
-			);
-			expect(results.length).toEqual(2);
-		});
-
-		it('should match ge', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('ge', 'field1 value 1')
-			);
-			expect(results.length).toEqual(2);
-		});
-
-		it('should match lt', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('lt', 'field1 value 2')
-			);
-			expect(results.length).toEqual(2);
-		});
-
-		it('should match le', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('le', 'field1 value 1')
-			);
-			expect(results.length).toEqual(2);
-		});
-
-		it('should match eq', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('eq', 'field1 value 1')
-			);
-			expect(results.length).toEqual(1);
-		});
-
-		it('should match ne', async () => {
-			const results = await DataStore.query(Model, m =>
-				m.field1('ne', 'field1 value 1')
-			);
-			expect(results.length).toEqual(2);
-		});
 	});
 
 	describe('Delete', () => {
@@ -347,136 +188,6 @@ describe('IndexedDBAdapter tests', () => {
 			// both should be undefined, even though we only explicitly deleted the user
 			expect(user).toBeUndefined;
 			expect(profile).toBeUndefined;
-		});
-	});
-
-	describe('Save', () => {
-		let User: PersistentModelConstructor<User>;
-		let Profile: PersistentModelConstructor<Profile>;
-		let Comment: PersistentModelConstructor<Comment>;
-		let Post: PersistentModelConstructor<Post>;
-		let adapter: any;
-
-		async function getMutations() {
-			await pause(250);
-			return await adapter.getAll('sync_MutationEvent');
-		}
-
-		beforeEach(async () => {
-			({ initSchema, DataStore } = require('../src/datastore/datastore'));
-
-			DataStore.configure({
-				storageAdapter: Adapter,
-			});
-			(DataStore as any).amplifyConfig.aws_appsync_graphqlEndpoint =
-				'https://0.0.0.0/does/not/exist/graphql';
-
-			const classes = initSchema(testSchema());
-
-			({ User, Profile, Comment, Post } = classes as {
-				User: PersistentModelConstructor<User>;
-				Profile: PersistentModelConstructor<Profile>;
-				Comment: PersistentModelConstructor<Comment>;
-				Post: PersistentModelConstructor<Post>;
-			});
-
-			await DataStore.clear();
-
-			// ensure `.storageAdapter` is set.
-			await DataStore.start();
-
-			adapter = (DataStore as any).storageAdapter;
-			const syncEngine = (DataStore as any).sync;
-
-			// my jest spy-fu wasn't up to snuff here. but, this succesfully
-			// prevents the mutation process from clearing the mutation queue, which
-			// allows us to observe the state of mutations.
-			(syncEngine as any).mutationsProcessor.isReady = () => false;
-		});
-
-		it('should allow linking model via model field', async () => {
-			const profile = await DataStore.save(
-				new Profile({ firstName: 'Rick', lastName: 'Bob' })
-			);
-
-			const savedUser = await DataStore.save(
-				new User({ name: 'test', profile })
-			);
-			const user1Id = savedUser.id;
-
-			const user = await DataStore.query(User, user1Id);
-			expect(user.profileID).toEqual(profile.id);
-			expect(user.profile).toEqual(profile);
-		});
-
-		it('should allow linking model via FK', async () => {
-			const profile = await DataStore.save(
-				new Profile({ firstName: 'Rick', lastName: 'Bob' })
-			);
-
-			const savedUser = await DataStore.save(
-				new User({ name: 'test', profileID: profile.id })
-			);
-			const user1Id = savedUser.id;
-
-			const user = await DataStore.query(User, user1Id);
-			expect(user.profileID).toEqual(profile.id);
-			expect(user.profile).toEqual(profile);
-		});
-
-		it('should produce a single mutation for an updated model with a BelongTo (regression test)', async () => {
-			// SQLite adapter, for example, was producing an extra mutation
-			// in this scenario.
-
-			const post = await DataStore.save(
-				new Post({
-					title: 'some post',
-				})
-			);
-
-			const comment = await DataStore.save(
-				new Comment({
-					content: 'some comment',
-					post,
-				})
-			);
-
-			const updatedComment = await DataStore.save(
-				Comment.copyOf(comment, draft => {
-					draft.content = 'updated content';
-				})
-			);
-
-			const mutations = await getMutations();
-
-			// comment update should be smashed to together with post
-			expect(mutations.length).toBe(2);
-			expectMutation(mutations[0], { title: 'some post' });
-			expectMutation(mutations[1], {
-				content: 'updated content',
-				postId: mutations[0].modelId,
-			});
-		});
-
-		it('should produce a mutation for a nested BELONGS_TO insert', async () => {
-			const comment = await DataStore.save(
-				new Comment({
-					content: 'newly created comment',
-					post: new Post({
-						title: 'newly created post',
-					}),
-				})
-			);
-
-			const mutations = await getMutations();
-
-			// one for the new comment, one for the new post
-			expect(mutations.length).toBe(2);
-			expectMutation(mutations[0], { title: 'newly created post' });
-			expectMutation(mutations[1], {
-				content: 'newly created comment',
-				postId: mutations[0].modelId,
-			});
 		});
 	});
 });
