+++ packages/datastore/__tests__/commonAdapterTests.ts
@@ -1,0 +1,347 @@
+import {
+	DataStore as DataStoreType,
+	PersistentModelConstructor,
+	initSchema as initSchemaType,
+} from '../src/';
+
+import {
+	pause,
+	expectMutation,
+	Model,
+	User,
+	Profile,
+	Post,
+	Comment,
+	testSchema,
+} from './helpers';
+
+export { pause };
+
+/**
+ * Adds common query test cases that all adapters should support.
+ *
+ * @param ctx A context object that provides a DataStore property, which returns
+ * a DataStore instance loaded with the storage adapter to test.
+ */
+export function addCommonQueryTests({
+	initSchema,
+	DataStore,
+	storageAdapter,
+	getMutations,
+}) {
+	describe('Common `query()` cases', () => {
+		let Model: PersistentModelConstructor<Model>;
+		let Comment: PersistentModelConstructor<Comment>;
+		let Post: PersistentModelConstructor<Post>;
+
+		beforeEach(async () => {
+			// ({ initSchema, DataStore } = require('../src/'));
+			DataStore.configure({ storageAdapter });
+
+			// establishing a fake appsync endpoint tricks DataStore into attempting
+			// sync operations, which we'll leverage to monitor how DataStore manages
+			// the outbox.
+			(DataStore as any).amplifyConfig.aws_appsync_graphqlEndpoint =
+				'https://0.0.0.0/does/not/exist/graphql';
+
+			const classes = initSchema(testSchema());
+			({ Comment, Model, Post } = classes as {
+				Comment: PersistentModelConstructor<Comment>;
+				Model: PersistentModelConstructor<Model>;
+				Post: PersistentModelConstructor<Post>;
+			});
+			await DataStore.clear();
+
+			// start() ensures storageAdapter is set
+			await DataStore.start();
+
+			const adapter = (DataStore as any).storageAdapter;
+			const db = (adapter as any).db;
+			const syncEngine = (DataStore as any).sync;
+
+			// my jest spy-fu wasn't up to snuff here. but, this succesfully
+			// prevents the mutation process from clearing the mutation queue, which
+			// allows us to observe the state of mutations.
+			(syncEngine as any).mutationsProcessor.isReady = () => false;
+
+			// NOTE: sort() test on these models can be flaky unless we
+			// strictly control the datestring of each! In a non-negligible percentage
+			// of test runs on a reasonably fast machine, DataStore.save() seemed to return
+			// quickly enough that dates were colliding. (or so it seemed!)
+
+			const baseDate = new Date();
+
+			await DataStore.save(
+				new Model({
+					field1: 'field1 value 0',
+					dateCreated: baseDate.toISOString(),
+				})
+			);
+			await DataStore.save(
+				new Model({
+					field1: 'field1 value 1',
+					dateCreated: new Date(baseDate.getTime() + 1).toISOString(),
+				})
+			);
+			await DataStore.save(
+				new Model({
+					field1: 'field1 value 2',
+					dateCreated: new Date(baseDate.getTime() + 2).toISOString(),
+				})
+			);
+		});
+
+		afterAll(async () => {
+			await DataStore.clear();
+
+			// prevent cross-contamination with other test suites that are not ~literally~
+			// expecting sync call counts to be ZERO.
+			(DataStore as any).amplifyConfig.aws_appsync_graphqlEndpoint = '';
+		});
+
+		it('should match fields of any non-empty value for `("ne", undefined)`', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('ne', undefined)
+			);
+			expect(results.length).toEqual(3);
+		});
+
+		it('should match fields of any non-empty value for `("ne", null)`', async () => {
+			const results = await DataStore.query(Model, m => m.field1('ne', null));
+			expect(results.length).toEqual(3);
+		});
+
+		it('should NOT match fields of any non-empty value for `("eq", undefined)`', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('eq', undefined)
+			);
+			expect(results.length).toEqual(0);
+		});
+
+		it('should NOT match fields of any non-empty value for `("eq", null)`', async () => {
+			const results = await DataStore.query(Model, m => m.field1('eq', null));
+			expect(results.length).toEqual(0);
+		});
+
+		it('should NOT match fields of any non-empty value for `("gt", null)`', async () => {
+			const results = await DataStore.query(Model, m => m.field1('gt', null));
+			expect(results.length).toEqual(0);
+		});
+
+		it('should NOT  match fields of any non-empty value for `("ge", null)`', async () => {
+			const results = await DataStore.query(Model, m => m.field1('ge', null));
+			expect(results.length).toEqual(0);
+		});
+
+		it('should NOT match fields of any non-empty value for `("lt", null)`', async () => {
+			const results = await DataStore.query(Model, m => m.field1('lt', null));
+			expect(results.length).toEqual(0);
+		});
+
+		it('should NOT match fields of any non-empty value for `("le", null)`', async () => {
+			const results = await DataStore.query(Model, m => m.field1('le', null));
+			expect(results.length).toEqual(0);
+		});
+
+		it('should NOT match fields of any non-empty value for `("gt", undefined)`', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('gt', undefined)
+			);
+			expect(results.length).toEqual(0);
+		});
+
+		it('should NOT match fields of any non-empty value for `("ge", undefined)`', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('ge', undefined)
+			);
+			expect(results.length).toEqual(0);
+		});
+
+		it('should NOT match fields of any non-empty value for `("lt", undefined)`', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('lt', undefined)
+			);
+			expect(results.length).toEqual(0);
+		});
+
+		it('should NOT match fields of any non-empty value for `("le", undefined)`', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('le', undefined)
+			);
+			expect(results.length).toEqual(0);
+		});
+
+		it('should match gt', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('gt', 'field1 value 0')
+			);
+			expect(results.length).toEqual(2);
+		});
+
+		it('should match ge', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('ge', 'field1 value 1')
+			);
+			expect(results.length).toEqual(2);
+		});
+
+		it('should match lt', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('lt', 'field1 value 2')
+			);
+			expect(results.length).toEqual(2);
+		});
+
+		it('should match le', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('le', 'field1 value 1')
+			);
+			expect(results.length).toEqual(2);
+		});
+
+		it('should match eq', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('eq', 'field1 value 1')
+			);
+			expect(results.length).toEqual(1);
+		});
+
+		it('should match ne', async () => {
+			const results = await DataStore.query(Model, m =>
+				m.field1('ne', 'field1 value 1')
+			);
+			expect(results.length).toEqual(2);
+		});
+	});
+
+	describe('Common `save()` cases', () => {
+		let Comment: PersistentModelConstructor<Comment>;
+		let Post: PersistentModelConstructor<Post>;
+		let Profile: PersistentModelConstructor<Profile>;
+		let User: PersistentModelConstructor<User>;
+		let adapter: any;
+
+		beforeEach(async () => {
+			// ({ initSchema, DataStore } = require('../src/'));
+			DataStore.configure({ storageAdapter });
+
+			// establishing a fake appsync endpoint tricks DataStore into attempting
+			// sync operations, which we'll leverage to monitor how DataStore manages
+			// the outbox.
+			(DataStore as any).amplifyConfig.aws_appsync_graphqlEndpoint =
+				'https://0.0.0.0/does/not/exist/graphql';
+
+			const classes = initSchema(testSchema());
+			({ User, Profile, Comment, Post } = classes as {
+				Comment: PersistentModelConstructor<Comment>;
+				Model: PersistentModelConstructor<Model>;
+				Post: PersistentModelConstructor<Post>;
+				Profile: PersistentModelConstructor<Profile>;
+				User: PersistentModelConstructor<User>;
+			});
+			await DataStore.clear();
+
+			// start() ensures storageAdapter is set
+			await DataStore.start();
+
+			adapter = (DataStore as any).storageAdapter;
+			const db = (adapter as any).db;
+			const syncEngine = (DataStore as any).sync;
+
+			// my jest spy-fu wasn't up to snuff here. but, this succesfully
+			// prevents the mutation process from clearing the mutation queue, which
+			// allows us to observe the state of mutations.
+			(syncEngine as any).mutationsProcessor.isReady = () => false;
+		});
+
+		afterAll(async () => {
+			await DataStore.clear();
+			(DataStore as any).amplifyConfig.aws_appsync_graphqlEndpoint = '';
+		});
+
+		it('should allow linking model via model field', async () => {
+			const profile = await DataStore.save(
+				new Profile({ firstName: 'Rick', lastName: 'Bob' })
+			);
+
+			const savedUser = await DataStore.save(
+				new User({ name: 'test', profile })
+			);
+			const user1Id = savedUser.id;
+
+			const user = await DataStore.query(User, user1Id);
+			expect(user.profileID).toEqual(profile.id);
+			expect(user.profile).toEqual(profile);
+		});
+
+		it('should allow linking model via FK', async () => {
+			const profile = await DataStore.save(
+				new Profile({ firstName: 'Rick', lastName: 'Bob' })
+			);
+
+			const savedUser = await DataStore.save(
+				new User({ name: 'test', profileID: profile.id })
+			);
+			const user1Id = savedUser.id;
+
+			const user = await DataStore.query(User, user1Id);
+			expect(user.profileID).toEqual(profile.id);
+			expect(user.profile).toEqual(profile);
+		});
+
+		it('should produce a single mutation for an updated model with a BelongTo (regression test)', async () => {
+			// SQLite adapter, for example, was producing an extra mutation
+			// in this scenario.
+
+			const post = await DataStore.save(
+				new Post({
+					title: 'some post',
+				})
+			);
+
+			const comment = await DataStore.save(
+				new Comment({
+					content: 'some comment',
+					post,
+				})
+			);
+
+			const updatedComment = await DataStore.save(
+				Comment.copyOf(comment, draft => {
+					draft.content = 'updated content';
+				})
+			);
+
+			const mutations = await getMutations(adapter);
+
+			// comment update should be smashed to together with post
+			expect(mutations.length).toBe(2);
+			expectMutation(mutations[0], { title: 'some post' });
+			expectMutation(mutations[1], {
+				content: 'updated content',
+				postId: mutations[0].modelId,
+			});
+		});
+
+		it('should produce a mutation for a nested BELONGS_TO insert', async () => {
+			const comment = await DataStore.save(
+				new Comment({
+					content: 'newly created comment',
+					post: new Post({
+						title: 'newly created post',
+					}),
+				})
+			);
+
+			const mutations = await getMutations(adapter);
+
+			// one for the new comment, one for the new post
+			expect(mutations.length).toBe(2);
+			expectMutation(mutations[0], { title: 'newly created post' });
+			expectMutation(mutations[1], {
+				content: 'newly created comment',
+				postId: mutations[0].modelId,
+			});
+		});
+	});
+}
