+++ packages/datastore/package.json
@@ -85,7 +85,8 @@
 		"testPathIgnorePatterns": [
 			"__tests__/model.ts",
 			"__tests__/schema.ts",
-			"__tests__/helpers.ts"
+			"__tests__/helpers.ts",
+			"__tests__/commonAdapterTests.ts"
 		],
 		"moduleFileExtensions": [
 			"ts",
