+++ packages/datastore/src/storage/adapter/InMemoryStore.ts
@@ -12,7 +12,7 @@ export class InMemoryStore {
 	multiRemove = async (keys: string[], callback?) => {
 		keys.forEach(k => this.db.delete(k));
 
-		callback();
+		typeof callback === 'function' && callback();
 	};
 
 	multiSet = async (entries: string[][], callback?) => {
@@ -20,7 +20,7 @@ export class InMemoryStore {
 			this.setItem(key, value);
 		});
 
-		callback();
+		typeof callback === 'function' && callback();
 	};
 
 	setItem = async (key: string, value: string) => {
